third_party/blink/web_tests/fast/dom/javascript-url-exception-isolation.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Exceptions thrown in javascript URLs should not propagate to the main script.")

var jsTestIsAsync = true;

var subframe = document.createElement("iframe");
document.body.appendChild(subframe);

var caughtException = false;
function test(idx) {
    switch (idx) {
        case 0:
            // Runtime exception.
            try {
                subframe.src = 'javascript:throw 42';
            } catch(e) {
                caughtException = true;
            }
            shouldBeFalse('caughtException');
            break;
        case 1:
            // Compile-time exception.
            try {
                subframe.src = 'javascript:<html></html>';
            } catch(e) {
                caughtException = true;
            }
            shouldBeFalse('caughtException');
            break;
        case 2:
            finishJSTest();
            return;
    }
    // Wait for script to evaluate.
    setTimeout(() => {
        test(idx + 1);
    }, 0);
}

test(0);

</script>
</body>
</html>
